# public/vitalsync-workout-openapi.yaml
openapi: 3.1.0
info:
  title: VitalSync Workout API
  version: 1.0.0
  description: 최근 운동 조회, 종목별 이력 조회, 확정 운동 저장 API
security:
  - bearerAuth: []
servers:
  - url: https://vitalsync-sigma.vercel.app
paths:
  /api/gpt/workout/context:
    post:
      operationId: getWorkoutContext
      summary: 최근 운동 세션과 종목별 최신 기록 조회
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 5
      responses:
        "200":
          description: 최근 운동 기록
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Bearer 인증 누락 또는 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Firestore 조회 또는 서버 내부 오류
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/workouts/exercises/recent:
    get:
      operationId: getRecentExerciseHistory
      summary: 특정 운동 종목의 최근 중량·횟수·세트 조회
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 5
      responses:
        "200":
          description: 종목별 최근 기록
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "404":
          description: 해당 종목 기록 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/workouts/{workoutId}:
    get:
      operationId: getWorkoutById
      summary: workoutId로 운동 상세 조회
      parameters:
        - name: workoutId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 운동 상세 기록
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "404":
          description: 운동 기록 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/workouts:
    get:
      operationId: listWorkoutSessions
      summary: 운동 세션 목록 조회
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 5
      responses:
        "200":
          description: 운동 세션 목록
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/gpt/workout/confirmed:
    post:
      operationId: saveConfirmedWorkout
      summary: 사용자가 저장을 확인한 운동 기록 저장
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutSaveRequest"
      responses:
        "201":
          description: 새 운동 저장 완료
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "200":
          description: 동일 idempotency key 기록이 이미 존재하여 duplicate 반환
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: 확정되지 않았거나 잘못된 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Bearer 인증 누락 또는 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Firestore 저장 또는 서버 내부 오류
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: VitalSync API key
  schemas:
    StrengthSet:
      type: object
      required: [setNumber, reps]
      properties:
        setNumber:
          type: integer
        weightKg:
          type: [number, "null"]
        reps:
          type: [integer, "null"]
        machineBaseWeightKg:
          type: [number, "null"]
        machineBaseWeightEstimated:
          type: boolean
        addedWeightKg:
          type: [number, "null"]
        estimated:
          type: boolean
        notes:
          type: string
    StrengthExercise:
      type: object
      required: [exerciseName, sets]
      properties:
        exerciseName:
          type: string
        canonicalExerciseId:
          type: string
        sets:
          type: array
          items:
            $ref: "#/components/schemas/StrengthSet"
        notes:
          type: string
    CardioExercise:
      type: object
      required: [exerciseName]
      properties:
        exerciseName:
          type: string
        durationMin:
          type: [number, "null"]
        speedKmh:
          type: [number, "null"]
        distanceKm:
          type: [number, "null"]
        inclinePercent:
          type: [number, "null"]
        notes:
          type: string
    WorkoutSaveRequest:
      type: object
      required: [confirmedByUser, date]
      properties:
        confirmedByUser:
          type: boolean
          const: true
        workoutId:
          type: string
        idempotencyKey:
          type: string
        idempotencyRequestId:
          type: string
        date:
          type: string
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
        type:
          type: string
        category:
          type: string
          enum: [personal, PT]
          default: personal
        duration:
          type: [number, "null"]
        durationEstimated:
          type: boolean
        startedAt:
          type: [string, "null"]
        endedAt:
          type: [string, "null"]
        strengthExercises:
          type: array
          items:
            $ref: "#/components/schemas/StrengthExercise"
        cardioExercises:
          type: array
          items:
            $ref: "#/components/schemas/CardioExercise"
        caloriesBurned:
          type: [number, "null"]
        rawMessages:
          type: array
          items:
            type: string
        location:
          type: string
        notes:
          type: string
    ErrorResponse:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          const: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - DUPLICATE
                - METHOD_NOT_ALLOWED
                - INTERNAL_ERROR
            message:
              type: string